feat: add page request test
