+++ src/router/routes/customer.js
@@ -115,6 +115,25 @@ const customerRoutes = [{
       component: () => import(/* webpackChunkName */ '../../views/Dashboard/Customer/RequestTest')
     },
     {
+      path: '/customer/request-test/find-lab',
+      name: 'find-lab',
+      meta: {
+        pageHeader: 'Find a Lab',
+        breadcrumbs: [
+          { text: 'Request a Test', href: '/customer/request-test' },
+          // { text: 'Checkout', disabled: true },
+        ],
+        drawerButtons: [
+          { text: "Dashboard", disabled: false, route: { name: "customer-home" } },
+          { text: "Request Test", active: true, disabled: false, route: { name: "request-test" } },
+          { text: "Test Results", disabled: false, route: { name: "all-test-result" } },
+          { text: "Order History", disabled: false, route: { name: "order-history" } },
+          { text: "Documents", disabled: false, route: { name: "documents" } },
+        ]
+      },
+      component: () => import(/* webpackChunkName */ '../../views/Dashboard/Customer/RequestTest/FindLab')
+    },
+    {
       path: '/customer/request-test/checkout',
       name: 'request-test-checkout',
       meta: {
